更新zxing，更新关键代码
更新zxing到最新版本。更新了编码解码的核心代码。

+++ zxing/oned/CodaBarReader.h
@@ -17,17 +17,17 @@
  * limitations under the License.
  */
 
-#include "zxing/oned/OneDReader.h"
-#include "zxing/common/BitArray.h"
-#include "zxing/Result.h"
+#include <zxing/oned/OneDReader.h>
+#include <zxing/common/BitArray.h>
+#include <zxing/Result.h>
 
 namespace zxing {
 namespace oned {
 
 class CodaBarReader : public OneDReader {
 private:
-  static const int MAX_ACCEPTABLE;
-  static const int PADDING;
+  static const float MAX_ACCEPTABLE;
+  static const float PADDING;
 
   // Keep some instance variables to avoid reallocations
   std::string decodeRowResult;
